competitive_analysis_agent/requirements.txt
google-adk
google-cloud-aiplatform
python-dotenv
pandas
requests
beautifulsoup4
openpyxl
python-pptx
google-api-python-client
flask
gunicorn
streamlit
nest_asyncio
